requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
appdirs==1.4.4
apturl==0.5.2
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.4
blis==0.7.9
breadability==0.1.20
Brlapi==0.8.3
bs4==0.0.1
cached-property==1.5.2
cachetools==5.3.0
catalogue==2.0.8
cchardet==2.1.7
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
ci-info==0.3.0
click==8.0.3
cmake==3.26.3
colorama==0.4.4
command-not-found==0.3
confection==0.0.4
configobj==5.0.8
configparser==5.3.0
contourpy==1.0.7
cryptography==40.0.2
cupshelpers==1.0
cycler==0.11.0
cymem==2.0.7
datasets==2.12.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.17.1
duplicity==0.8.21
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
entrypoints==0.4
etelemetry==0.3.0
exceptiongroup==1.1.1
executing==1.2.0
fasteners==0.14.1
feedparser==6.0.8
filelock==3.11.0
fitz==0.0.1.dev2
flatbuffers==23.3.3
fonttools==4.39.3
freezegun==1.0.0
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.2
gast==0.4.0
gensim==4.3.1
gevent==21.8.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googletrans==4.0.0rc1
googletranslate==2.1.2
graphsurgeon @ file:///home/codenomad/Downloads/TensorRT-8.6.1.6/graphsurgeon/graphsurgeon-0.4.6-py2.py3-none-any.whl
greenlet==1.1.2
grpcio==1.54.0
gTTS==2.3.2
gyp==0.1
h11==0.9.0
h2==3.2.0
h5py==3.8.0
hpack==3.0.0
hstspreload==2023.1.1
html2text==2020.1.16
html5lib==1.1
httpcore==0.9.1
httplib2==0.20.2
httpx==0.13.3
huggingface-hub==0.14.1
hyperframe==5.2.0
idna==3.3
imageio==2.31.1
importlib-metadata==4.6.4
imutils==0.5.4
ipython==8.14.0
isodate==0.6.1
jax==0.4.8
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.12.0
keras-models==0.0.7
keyring==23.5.0
kiwisolver==1.4.4
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
libclang==16.0.0
librosa==0.10.0.post2
libsass==0.20.1
lit==16.0.1
llvmlite==0.40.1
lockfile==0.12.2
looseversion==1.1.2
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.1.0
mock==4.0.3
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
mysqlclient==2.1.1
netifaces==0.11.0
nibabel==5.1.0
nipype==1.8.6
nltk==3.8.1
num2words==0.5.10
numba==0.57.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
odoo==14.0
ofxparse==0.21
olefile==0.46
onnx==1.14.0
onnx-graphsurgeon @ file:///home/codenomad/Downloads/TensorRT-8.6.1.6/onnx_graphsurgeon/onnx_graphsurgeon-0.3.12-py2.py3-none-any.whl
openai==0.27.4
opencv-python==4.7.0.72
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.1
pandas==2.0.0
paramiko==2.9.3
parso==0.8.3
passlib==1.7.4
pathlib==1.0.1
pathtools==0.1.2
pathy==0.10.1
patsy==0.5.3
pbr==5.8.0
pdfminer.six==20221105
pdfplumber==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
playsound==1.3.0
polib==1.1.0
pooch==1.6.0
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==4.23.1
prov==2.0.0
psutil==5.9.5
psycopg2==2.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.11
pycairo==1.20.1
pycountry==22.3.5
pycparser==2.21
pycups==2.0.1
pydantic==1.10.7
pydeck==0.8.1b0
pydot==1.4.2
pydub==0.25.1
Pygments==2.15.1
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
Pympler==1.0.1
PyMuPDF==1.22.3
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pypdf==3.9.0
PyPDF2==3.0.1
pyRFC3339==1.1
pyrsistent==0.19.3
pyserial==3.5
PySocks==1.7.1
pytesseract==0.3.10
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
python-ldap==3.4.0
python-stdnum==1.16
python-version==0.0.2
pyttsx3==2.90
pytz==2022.1
pyusb==1.0.2
PyWavelets==1.4.1
pyxdg==0.27
pyxnat==1.5
PyYAML==5.4.1
qrcode==6.1
rdflib==6.3.2
regex==2023.3.23
reportlab==3.5.59
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
responses==0.18.0
rfc3986==1.5.0
rich==13.3.5
rjsmin==1.1.0
roman==3.3
rsa==4.9
scikit-fuzzy==0.4.2
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.11.0
screen-resolution-extra==0.0.0
seaborn==0.12.2
SecretStorage==3.3.1
selenium==4.9.0
sentencepiece==0.1.99
sentry-sdk==1.21.0
seqeval==1.2.2
setproctitle==1.3.2
sgmllib3k==1.0.0
simplejson==3.19.1
simpletransformers==0.63.11
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.4
srsly==2.4.6
stack-data==0.6.2
statsmodels==0.13.5
streamlit==1.22.0
stripe==5.4.0
summa==1.2.0
sumy==0.11.0
sympy==1.11.1
systemd-python==234
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorrt @ file:///home/codenomad/Downloads/TensorRT-8.6.1.6/python/tensorrt-8.6.1-cp310-none-linux_x86_64.whl
termcolor==2.3.0
thinc==8.1.9
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.4.12
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
traits==6.3.2
transformers==4.28.1
trio==0.22.0
trio-websocket==0.10.2
triton==2.0.0
typer==0.7.0
typing_extensions==4.6.0
tzdata==2023.3
tzlocal==5.0.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
uff @ file:///home/codenomad/Downloads/TensorRT-8.6.1.6/uff/uff-0.6.9-py2.py3-none-any.whl
ufw==0.36.1
ultralytics==8.0.89
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
validators==0.20.0
vboxapi==1.0
virtualenv==20.21.0
vobject==0.9.6.1
wadllib==1.3.6
Wand==0.6.11
wandb==0.15.2
wasabi==1.1.1
watchdog==3.0.0
wcwidth==0.2.6
webdriver-manager==3.8.6
webencodings==0.5.1
Werkzeug==2.3.3
wrapt==1.14.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
xxhash==3.2.0
yarl==1.9.1
zeep==4.1.0
zipp==1.0.0
zope.event==4.4
zope.interface==5.4.0
